requirements.txt
# ============================================================
# News Pipeline — Toutes les dépendances (Phases 1 à 4)
# ============================================================

# Phase 1 — Scraping & Data Lake local
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.1
pandas==2.2.2
pyarrow==16.0.0

# Phase 2 — Infrastructure distribuée (Docker)
kafka-python==2.0.2
minio==7.2.7
# apache-airflow==2.9.1   # Installé dans le conteneur Docker uniquement

# Phase 3 — Différenciateurs (BERTopic + Polymarket)
bertopic==0.16.2
sentence-transformers==2.7.0
# torch doit être installé séparément selon la plateforme :
# CPU only  : pip install torch --index-url https://download.pytorch.org/whl/cpu
# GPU (CUDA): pip install torch --index-url https://download.pytorch.org/whl/cu121

# Phase 4 — Dashboard
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0
